Fixed log4j implementation
Now using log4j2.xml, which will also serve as a starting point for logging.  In the future we can programatically change the verbosity levels if we want to honor the ones specified in the testng xml and therefore set via jenkins.

Tweaked the email reports and changed a lot of println statements to log4j.